Experimental analysis code.
Unused, just playing around.

Enough to find the module-level contents but doesn't attempt to type any
of them, so it'll be useless as soon as we get to microbit.display :)
Will add just enough typing and see how that goes. Will need to start
annotating the tree somehow though or it'll get really painful.

Intending to merge this with the "bottom up" completion code we already
have.

Still undecided as to whether it's worth a CST -> AST pass first.