Partial flashing.
Substantially the same as in v2.

Needs significant refactoring and conversion to TypeScript, but I think
it's in a working state and ought to have all the same bugs :)

I've somewhat gutted my top-level connection class as it'll be easier to
reimplement serial and connection states on top of this.