More typing for partial flashing.
The constants situation is a bit unfortunate. CRA via babel doesn't
support const enums (which are inlined) so I've not been able to use the
type definitions from pxt (or those that are in dapjs). When we extract
a separate library that could change so I've tried to keep the access
style source compatible.