Centralised logging.
Preparatory work for Sentry crash and GA integration (neither enabled
for now, both optional).

Holding off configuring this until we have a CircleCI build.

I've not reinstated any of the WebUSB events - we need to carefully
consider the adapter layer the v2 code has.

The events added to project-actions.ts are very simple. Need to consider
whether to align them. We should distinguish more cases for sure but
this is enough to be helpful as context for Sentry in initial testing.