Review and test file system (#36)
Strip out the file size stuff for now.
Strip out local storage for now but keep everything async so it's easy to add e.g. IndexDB in future.
Add an upload action. For sure the wrong location and behaviour, but useful for discussion.
Some stub test cases added to fill out as we resolve questions.

https://github.com/microbit-foundation/python-editor-next/issues/21