Left nav curve (#130)
Doesn't match the design, but we want to keep the text so will have to
tweak. "Placeholder" is too long so went with Jonny's suggestion.

Use an alternative focus style.

There seem to be issues with zoom and SVG positioning against the right
edge. Needs some more thought, but worth getting this in anyway.

https://github.com/microbit-foundation/python-editor-next/issues/120