Add a dev container (#134)
Uses node 14 and runs npm install.

Git config seems to be forwarded to the container but it drops the
editor settings, so configure VS Code as the editor. Seem to have to
do it when "started" for it to be late enough.

Let's see if this is a good option for Windows users.