Add translations using react-intl.
There are a few bits to fix still but this covers the bulk of the text.
Message ids still to be reviewed.
Feedback copy intentionally omitted.
No Crowdin integration yet (but the Chrome JSON format looks like it'll
do the trick).

npm run i18n:compile

is the new command to compile messages.