Optionally use external project for theme etc. (#151)
This commit introduces some indirection via a webpack alias to allow us to have a single import of @deployment that is switched at build time to use an external module containing brand assets or an internal default one.

We use https://github.com/gsoft-inc/craco to allow webpack config customisation as CRA doesn't permit it by default. This carries some risk of slower CRA upgrades in future but I think we need something like this to get a plugin-like feature.

I've gone with an import directly from a "scope" as it can't really be confused with a real NPM package.