Move collapsed behaviour to the split pane.
This is working pretty well now.
The cursor is a bit awkward when resizing, as the two text areas either
side cause an issue. Not sure what to do really. We could overlay them
when dragging.

This could really do with storybook now but I'm reluctant to add it due
to the inevitable future build/upgrade churn.