Code structure highlighting: collapse implementations before adding more config (#209)
This removes the "brackets" implementation that we've decided against progressing and collapses the code that implements the remaining approaches into a single parameterised implementation. We lean more heavily on CSS, but there is a fundamental difference between l-shapes and boxes that needs switching in the view code.

The new parameters are not yet exposed to the user. Combinations of parameters that aren't surfaced through the settings will need careful review after this PR.

Folks who previously had brackets enabled will just get no highlighting for now (probably no-one).

Various tidying / splitting up of the code.

See #207

I've been testing this with the default theme and it also fixes #204.