Vary highlighting based on cursor position.
Feels buggy at the end of blocks due to disagreements about where the
block ends. We need to review and discuss this.

Behaviour is also weird at the end of the file as the last position
isn't considered inside the last Body node (see equivalent issue in
indentation logic). We can add special logic for that case.