Make sure we have only the most nested block.
It's still confusing as cursor positions in the indent area are part of
the block (sort of). Same is true of the cursor positions around the
colon where the body starts.

Remove the whitespace trimming as it's more confusing now we've added
highlighting. On reflection I think having it is better for code entry.