Basic TygerPython linting.
CM's defaults.

Presumably needs moving to a worker to perform well on larger documents,
but this is a start.

Bundle size needs checking. ~700k is their usual build but to get it
importing I've gone with the ESM one. Not clear if we'll mimimise as
well as their Closure Compiler version does for their main build.