Beginnings of a collapsed presentation of serial.
Expand/collapse behaviour not yet implemented. It's complicated because
the terminal seems to need a known pixel height and we already have
the split pane.

Not at all sure the error message belongs here but I'd like to work it
up as an option, and I think we want the collapse feature anyway. I'll
also look at inline presentations.