Show runtime errors in compact serial view (#224)
- Add serial utility features from V2 for Ctrl-D, Ctrl-C. 
- Add some minimal in-app help to explain what's going on.

We've stayed away from inline presentation of errors in the editor as runtime errors naturally get out of sync with the current code and we feel it'll be more confusing then helpful. Work is planned for lint errors that will be shown inline as they'll be up-to-date.

There's little by way of e2e testing here. I've made a start on #226 as we'll want to mock the serial interactions.
I really want to cover the line/file navigation.