Serial mock e2e tests (#244)
We dispatch a CustomEvent from Puppeteer which we then emit as a serial_data event.

Allows us to test the terminal and serial area.

More tests to follow. I want to add primitives to let us test the open file and line number so I can cover the traceback navigation.