Avoid disconnect during flash. (#252)
Somewhat speculative fix as I can't get the visibilitychange event to trigger during flash anymore. It seems to occur at the end of flash. But given there's async work in the flash it makes sense that it can occur.

I've added logging for the visibility change work.

I think we'd like to drop all this if we could fix the underlying cause of #89.

Fixes https://github.com/microbit-foundation/python-editor-next/issues/223