Signature help (#272)
Initial exploration of signature help.

Behind a flag for now. It will only be enabled on review branches.

I'm particularly interested in using it with the apidocs branches.

The tooltip that pops up is inadequate, just displaying a formatted docstring. I'm going to look at adding documentation during autocomplete and then address the presentation of those together.