Add "Advanced" API documentation section to sidebar (feature flagged) (#269)
This relies on a custom LSP extension to pull the relevant data from Pyright (available on the microbit branch of https://github.com/microbit-matt-hillsdon/pyright/ – it'll move to the microbit-foundation org before long.

It also relies on a round of editing to the stubs, reflected in the new typeshed. This edit is not yet final but is much better aligned with what we want from them, with standalone first lines of docstrings.

The design will change to align with in-progress design work for more beginner-oriented documentation. We won't remove the feature flag until we have that documentation ready to go live.

See https://github.com/microbit-foundation/python-editor-next/issues/263