Add the imports we'll need to the mock data.
Strip them out for display.
We now have the right input for the edit calculation, it's just a matter
of communicating in a sane way between the CM code and the React
component. We probably need a ref to the active editor held further up
the tree, then we can have an "active editor actions" class or similar.