Show "ghost code" during drag and drop (#418)
We support only a single drag or drop, i.e. dragging again will clear the existing decorations. The highlighted text can be non-contiguous though (consider imports).

We could consider changing this but I think it'd be more confusing than helpful to have multigenerational ghosts.

Colour is non-standard but likely to remain specific to this unusual use so I've left it out of the theme.

The drop highlighting also works for "Insert code" but is held back by the same dnd flag for now.