Switch away from homepage to PUBLIC_URL.
Set the url to the actual path. Motivated by an issue with dynamic
import in a worker.

Use alternative to PUBLIC_URL for e2e.

CRA's init process seems to clear PUBLIC_URL which is kinda unhelpful
when we want to use it in e2e. Happens even with a trivial jest test.