Files header with explanatory text (#604)
Going with this for user testing tomorrow, but I definitely think it is worth revisiting. 

It'd be nice to make the files list clearer. It used to have a white background on a darker grey but we changed the grey across all the sidebar tabs to something with too little contrast with the white.

I'd also prefer it if sidebar buttons had a more distinctive style from those in the bottom bar.

Closes #585