Files -> Project tab (#664)
Use the project name as a heading to make it clearer that the files
belong to the project you're editing, rather than each file being a
project. This addresses user testing feedback.

Prefer "Add" terminology to "Create" to make it clearer the file is
added to the project, not creating anything new.