Add error markers in the gutter with a friendlier rendering while the line is being edited (#634)
Requires us to inline CM's MIT licensed lint extension.

Once the dust has settled from this feature we'll do some work to see if we can propose a way of reusing more of CM's code here. There isn't an obvious seam between the gutter and the rest of the lint module, but perhaps one can be found. I think we may change this significantly so we should wait a few weeks.

Co-authored-by: Matt Hillsdon <matt.hillsdon@microbit.org>