Add WebUSB assistance dialogs (#702)
Provides clearer more persistent information to users about the first connection and if they need to upgrade firmware.

Disabled for non-English (French, in practice) languages due to the large number of untranslated strings we'd be requiring the user to understand for a key usecase. We'll reenable after the next round of translation or if we move to community translation.