Improve common student error messages (#747)
Updates Pyright to https://github.com/microbit-foundation/pyright/commit/2732ac85a90ae28be0cead8ead0d6e1a8990fd12

This has the first pass of identifying a subset of error messages to simplify for our audience, and a larger set that we intend to translate. Further work to follow based on feedback on the simplifications and to support translation.