"Add file" -> "Create file" (#771)
We changed it from "Create new file" to "Add file" on #664 along with other changes to make it clearer that it's a file in your project not a way to create a new project. Based on support interactions and other feedback folks are now using it with the expectation it's the route to upload a file to a project. On balance we think it's still confusing, and hope that the other changes and just "Create file" make it clearer. 

Closes https://github.com/microbit-foundation/python-editor-next/issues/761