"Send to micro:bit" primary action instead of Connect then Flash (#763)
Based on user feedback, the Flash terminology is not well understood.

This PR combines Connect/Flash in a single action.

If you really want to Connect without Flash then you can do so via the "⋮" menu, likewise Disconnect.

It has the additional benefit that "Download" is always visible for users to save their work.

This temporarily leaks English into the French translation in a key location. This will be addressed for the live release.