Minimal while smooshing (#799)
"While smooshing" avoids the nested while True constructs that are
easy to create from the simple examples available to drag and drop.

This doesn't cope with while loops with a preamble (see e.g. variables
section). For that we'd need to split the preamble which would add a
fair amount of complexity. It copes well with the examples in Buttons
and Accelerometer though. We may revisit this in future.

We also considered highlighting the while True in yellow and may revisit
that too.