Link to Reference from autocomplete and parameter help. (#806)
Previously we linked to API from a link titled "Help".

We now link to Reference from a link titled "Help" and to API from a link titled "API".

For the moment, either link can be absent, as we're missing API documentation for builtins. Not every API term has relevant Reference documentation.

In future, once we've documented the builtins module, API will always be present.