Update MicroPython for V2.
New version is from
https://github.com/microbit-foundation/micropython-microbit-v2/actions/runs/2774388121
with the fix for https://github.com/microbit-foundation/micropython-microbit-v2/issues/101

There are no other changes on master so this is just a bug fix.

The SoundEffect work remains on a branch. We'll live with the
audioSoundEffect branch/flag having broken neopixels until that's merged
in the next few weeks.