webusb: On serial connection check baud rate before setting it.
Setting the baud rate makes DAPLink send a "break signal", which
resets the target.

This is triggered  when switching tabs as it resets the user's
micro:bits when the editor goes out of focus and active again..

We can avoid this reset by checking first the DAPLink serial
settings and only set the baud rate if necessary.