Ensure we don't index when language doesn't match
Previously we pulled the language off what we indexed which could cause
an issue when the search index was updated for a new language but the
content had not been reloaded. Instead an index now knows its language.

Additionally, ensure we wait for the languages of all content to match
before indexing.