Signature help: restructure async work (#1198)
sendRequest isn't an async function and it can synchronously throw errors
so we need something to prevent reentrant updates in the error case.

Otherwise though we don't needed it if we move more work into the state.

I've also calculated line/column ASAP - I'm not sure this is necessary but it
feels obviously correct / straightforward which is a win in this code.